Numeros reais e inteiros sendo analisados em compile time

+++ src/expressions/AbstractExpression.java
@@ -1,7 +1,7 @@
 package expressions;
 
 public abstract class AbstractExpression {
-	
-	public abstract int eval();	
+
+	public abstract double eval();
 
 }
